+++ src/components/Accordion/accordion.scss
@@ -1,4 +1,4 @@
-@import '../assets/scss/variables';
+@import '../../assets/scss/variables';
 
 .inactive {
     display: none;
