+++ src/components/Navbar/navbar.scss
@@ -1,4 +1,4 @@
-@import '../assets/scss/variables';
+@import '../../assets/scss/variables';
 
 .navigation {
     align-items: center;
