+++ src/components/Tabs/Tabs.scss
@@ -1,4 +1,4 @@
-@import '../assets/scss/variables';
+@import '../../assets/scss/variables';
 
 .inactive{
     display: none;
